requirements-lock.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements-lock.txt pyproject.toml
#
aioquic==1.2.0
    # via mitmproxy
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
asgiref==3.8.1
    # via mitmproxy
attrs==25.4.0
    # via service-identity
blinker==1.9.0
    # via flask
brotli==1.1.0
    # via mitmproxy
cachetools==5.5.2
    # via agentic-chaos-monkey (pyproject.toml)
certifi==2026.1.4
    # via
    #   aioquic
    #   httpcore
    #   httpx
    #   mitmproxy
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   typer
    #   uvicorn
cryptography==43.0.3
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   aioquic
    #   mitmproxy
    #   pyopenssl
    #   service-identity
fastapi==0.128.0
    # via agentic-chaos-monkey (pyproject.toml)
flask==3.0.3
    # via mitmproxy
googleapis-common-protos==1.72.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.76.0
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.14.0
    # via
    #   httpcore
    #   mitmproxy
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via mitmproxy
hpack==4.1.0
    # via h2
httpcore==1.0.8
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   langsmith
    #   ollama
hyperframe==6.0.1
    # via
    #   h2
    #   mitmproxy
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jsonpatch==1.33
    # via langchain-core
jsonpath-ng==1.7.0
    # via agentic-chaos-monkey (pyproject.toml)
jsonpointer==3.0.0
    # via jsonpatch
kaitaistruct==0.10
    # via mitmproxy
langchain-core==0.2.43
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   langchain-ollama
langchain-ollama==0.1.3
    # via agentic-chaos-monkey (pyproject.toml)
langsmith==0.1.147
    # via langchain-core
ldap3==2.9.1
    # via mitmproxy
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
mitmproxy==10.4.2
    # via agentic-chaos-monkey (pyproject.toml)
mitmproxy-macos==0.6.3
    # via mitmproxy-rs
mitmproxy-rs==0.6.3
    # via mitmproxy
msgpack==1.0.8
    # via mitmproxy
ollama==0.6.1
    # via langchain-ollama
opentelemetry-api==1.39.1
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-exporter-prometheus
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.39.1
    # via agentic-chaos-monkey (pyproject.toml)
opentelemetry-exporter-otlp-proto-common==1.39.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.39.1
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.39.1
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus==0.60b1
    # via agentic-chaos-monkey (pyproject.toml)
opentelemetry-instrumentation==0.60b1
    # via opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-httpx==0.60b1
    # via agentic-chaos-monkey (pyproject.toml)
opentelemetry-proto==1.39.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.39.1
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-exporter-prometheus
opentelemetry-semantic-conventions==0.60b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
opentelemetry-util-http==0.60b1
    # via opentelemetry-instrumentation-httpx
orjson==3.11.5
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   opentelemetry-instrumentation
passlib==1.7.4
    # via mitmproxy
ply==3.11
    # via jsonpath-ng
prometheus-client==0.24.1
    # via opentelemetry-exporter-prometheus
protobuf==5.27.3
    # via
    #   googleapis-common-protos
    #   mitmproxy
    #   opentelemetry-proto
publicsuffix2==2.20191221
    # via mitmproxy
pyasn1==0.6.2
    # via
    #   ldap3
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   fastapi
    #   langchain-core
    #   langsmith
    #   ollama
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via agentic-chaos-monkey (pyproject.toml)
pylsqpack==0.3.23
    # via aioquic
pyopenssl==24.2.1
    # via
    #   aioquic
    #   mitmproxy
pyparsing==3.1.2
    # via mitmproxy
pyperclip==1.9.0
    # via mitmproxy
python-dotenv==1.2.1
    # via uvicorn
pyyaml==6.0.3
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   langchain-core
    #   uvicorn
requests==2.32.5
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   langsmith
    #   opentelemetry-exporter-otlp-proto-http
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.4
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   typer
ruamel-yaml==0.18.6
    # via mitmproxy
ruamel-yaml-clib==0.2.15
    # via ruamel-yaml
service-identity==24.2.0
    # via aioquic
shellingham==1.5.4
    # via typer
sortedcontainers==2.4.0
    # via mitmproxy
starlette==0.50.0
    # via fastapi
tenacity==8.5.0
    # via
    #   agentic-chaos-monkey (pyproject.toml)
    #   langchain-core
tornado==6.4.1
    # via mitmproxy
typer==0.21.1
    # via agentic-chaos-monkey (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   grpcio
    #   langchain-core
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typer
    #   typing-inspection
    #   urwid
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
urwid==2.6.15
    # via mitmproxy
uvicorn[standard]==0.40.0
    # via agentic-chaos-monkey (pyproject.toml)
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.2.14
    # via urwid
websockets==16.0
    # via uvicorn
werkzeug==3.1.5
    # via flask
wrapt==1.17.3
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
wsproto==1.2.0
    # via mitmproxy
zipp==3.23.0
    # via importlib-metadata
zstandard==0.23.0
    # via mitmproxy
